Working and tested version of the app
