+++ app/src/main/java/com/lastaoutdoor/lasta/data/api/ApiService.kt
@@ -1,8 +1,8 @@
 package com.lastaoutdoor.lasta.data.api
 
-import com.lastaoutdoor.lasta.data.model.Node
-import com.lastaoutdoor.lasta.data.model.Relation
-import com.lastaoutdoor.lasta.data.model.Way
+import com.lastaoutdoor.lasta.data.model.api.Node
+import com.lastaoutdoor.lasta.data.model.api.Relation
+import com.lastaoutdoor.lasta.data.model.api.Way
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
