--- app/src/main/java/com/lastaoutdoor/lasta/data/model/TimeFrame.kt
@@ -1,7 +1,0 @@
-package com.lastaoutdoor.lasta.data.model
-enum class TimeFrame {
-  W,
-  M,
-  Y,
-  ALL
-}
