+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/activity/Difficulty.kt
@@ -1,8 +1,9 @@
-package com.lastaoutdoor.lasta.data.model
+package com.lastaoutdoor.lasta.data.model.activity
 
-enum class Sports {
-  HIKING,
-  CLIMBING;
+enum class Difficulty {
+  EASY,
+  MODERATE,
+  DIFFICULT;
 
   override fun toString(): String {
     return name.lowercase().replaceFirstChar { it.uppercase() }
