+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/activity/OutdoorActivity.kt
@@ -1,4 +1,6 @@
-package com.lastaoutdoor.lasta.data.model
+package com.lastaoutdoor.lasta.data.model.activity
+
+import com.lastaoutdoor.lasta.data.model.activity.ActivityType
 
 open class OutdoorActivity(
     private var activityType: ActivityType,
