+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/api/Relation.kt
@@ -1,7 +1,9 @@
-package com.lastaoutdoor.lasta.data.model
+package com.lastaoutdoor.lasta.data.model.api
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import com.lastaoutdoor.lasta.data.model.activity.ActivityType
+import com.lastaoutdoor.lasta.data.model.activity.OutdoorActivity
 
 class Relation(
     @SerializedName("type") @Expose val type: String,
