+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/profile/Popularity.kt
@@ -1,3 +1,3 @@
-package com.lastaoutdoor.lasta.data.model
+package com.lastaoutdoor.lasta.data.model.profile
 
 data class Popularity(var rating: Float = 0f, var numRatings: Int = 0)
