+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/profile/TimeFrame.kt
@@ -1,0 +1,7 @@
+package com.lastaoutdoor.lasta.data.model.profile
+enum class TimeFrame {
+  W,
+  M,
+  Y,
+  ALL
+}
