+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/user/HikingLevel.kt
@@ -1,0 +1,8 @@
+package com.lastaoutdoor.lasta.data.model.user
+
+/** Enum class for hiking levels */
+enum class HikingLevel(val level: Int) {
+    BEGINNER(0),
+    INTERMEDIATE(1),
+    ADVANCED(2)
+}
