+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/user/UserPreferences.kt
@@ -1,0 +1,10 @@
+package com.lastaoutdoor.lasta.data.model.user
+
+data class UserPreferences(
+    val isLoggedIn: Boolean,
+    val uid: String,
+    val userName: String,
+    val email: String,
+    val profilePictureUrl: String,
+    val hikingLevel: HikingLevel
+)
