+++ app/src/main/java/com/lastaoutdoor/lasta/repository/OutdoorActivityRepository.kt
@@ -2,10 +2,10 @@ package com.lastaoutdoor.lasta.repository
 
 import com.lastaoutdoor.lasta.data.api.ApiService
 import com.lastaoutdoor.lasta.data.api.OutdoorActivityResponse
-import com.lastaoutdoor.lasta.data.model.ActivityType
-import com.lastaoutdoor.lasta.data.model.Node
-import com.lastaoutdoor.lasta.data.model.Relation
-import com.lastaoutdoor.lasta.data.model.Way
+import com.lastaoutdoor.lasta.data.model.activity.ActivityType
+import com.lastaoutdoor.lasta.data.model.api.Node
+import com.lastaoutdoor.lasta.data.model.api.Relation
+import com.lastaoutdoor.lasta.data.model.api.Way
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
